requirements.txt
annotated-types              0.6.0      
anyio                        4.3.0      
APScheduler                  3.10.4     
asn1crypto                   1.5.1      
cachetools                   5.3.3      
certifi                      2024.2.2   
charset-normalizer           3.3.2      
click                        8.1.7      
colorama                     0.4.6      
distro                       1.9.0      
dnspython                    2.6.1      
email_validator              2.1.1
et-xmlfile                   1.1.0
exceptiongroup               1.2.1
Faker                        25.3.0
fastapi                      0.111.0
fastapi-cli                  0.0.3
google-ai-generativelanguage 0.6.4
google-api-core              2.19.0
google-api-python-client     2.129.0
google-auth                  2.29.0
google-auth-httplib2         0.2.0
google-generativeai          0.5.4
googleapis-common-protos     1.63.0
greenlet                     3.0.3
grpcio                       1.63.0
grpcio-status                1.62.2
h11                          0.14.0
httpcore                     1.0.5
httplib2                     0.22.0
httptools                    0.6.1
httpx                        0.27.0
idna                         3.7
Jinja2                       3.1.4
markdown-it-py               3.0.0
MarkupSafe                   2.1.5
mdurl                        0.1.2
numpy                        1.26.4
openai                       1.30.1
openpyxl                     3.1.2
orjson                       3.10.3
pandas                       2.2.2
pg8000                       1.31.2
pip                          22.0.4
proto-plus                   1.23.0
protobuf                     4.25.3
psycopg2                     2.9.9
psycopg2-binary              2.9.9
pyasn1                       0.6.0
pyasn1_modules               0.4.0
pydantic                     2.7.1
pydantic_core                2.18.2
Pygments                     2.18.0
pymongo                      4.7.2
pyparsing                    3.1.2
python-dateutil              2.9.0.post0
python-dotenv                1.0.1
python-multipart             0.0.9
pytz                         2024.1
PyYAML                       6.0.1
requests                     2.31.0
rich                         13.7.1
rsa                          4.9
scramp                       1.4.5
setuptools                   58.1.0
shellingham                  1.5.4
six                          1.16.0
sniffio                      1.3.1
SQLAlchemy                   2.0.30
starlette                    0.37.2
tqdm                         4.66.4
typer                        0.12.3
typing_extensions            4.11.0
tzdata                       2024.1
tzlocal                      5.2
ujson                        5.10.0
uritemplate                  4.1.1
urllib3                      2.2.1
uvicorn                      0.29.0
watchfiles                   0.21.0
websockets                   12.0